Revert "update gst-plugins-bad"
This reverts commit eb06725b7edbb1c29c57e0e276a7f8488e644f1f.

+++ gst-plugins-bad/ext/libmms/gstmms.h
@@ -32,7 +32,7 @@ struct _GstMMS
 
   gchar  *uri_name;
   gchar  *current_connection_uri_name;
-  guint64  connection_speed;
+  guint  connection_speed;
   
   mmsx_t *connection;
 };
